% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pli-course-api.R
\name{pl_assessment_instance_list}
\alias{pl_assessment_instance_list}
\alias{pl_assessment_instance_view}
\title{View Assessment Instances}
\usage{
pl_assessment_instance_list(course_instance_id, assessment_id)

pl_assessment_instance_view(course_instance_id, assessment_instance_id)
}
\arguments{
\item{course_instance_id}{Integer value associated with the course.}

\item{assessment_id}{Integer of a specific assessment to view.}

\item{assessment_instance_id}{Integer of student assessment.}
}
\description{
See a list of all assessment instances for a given assessment or view
results from a single assessment instance.
}
