% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pli-course-api.R
\name{pl_assessment_submission_list}
\alias{pl_assessment_submission_list}
\alias{pl_assessment_submission_view}
\title{View Submissions on Assessment Instances}
\usage{
pl_assessment_submission_list(course_instance_id, assessment_instance_id)

pl_assessment_submission_view(course_instance_id, submission_id)
}
\arguments{
\item{course_instance_id}{Integer value associated with the course.}

\item{assessment_instance_id}{Integer of student assessment.}

\item{submission_id}{Integer of student submission attempt.}
}
\description{
Obtain a list of all submissions within a given assessment instances or view
results from a single submission inside the course.
}
