% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pli-course-api.R
\name{pl_assessment_list}
\alias{pl_assessment_list}
\alias{pl_assessment_view}
\title{View Assessment Overviews}
\usage{
pl_assessment_list(course_instance_id)

pl_assessment_view(course_instance_id, assessment_id)
}
\arguments{
\item{course_instance_id}{Integer value associated with the course.}

\item{assessment_id}{Integer of a specific assessment to view.}
}
\value{
A \code{data.frame}
}
\description{
See a list of all assessments or view data for a single assessment.
}
