% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lms-export.R
\name{grade_format_compass2g}
\alias{grade_format_compass2g}
\title{Export grades to Compass2g Format}
\usage{
grade_format_compass2g(x, path, assignment = "Points")
}
\arguments{
\item{x}{Data set}

\item{path}{Name and path to save CSV.}

\item{assignment}{Pull point data}
}
\description{
Export grades to Compass2g Format
}
