% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pli-course-api.R
\name{pl_course_gradebook}
\alias{pl_course_gradebook}
\title{Course gradebook}
\usage{
pl_course_gradebook(course_instance_id)
}
\arguments{
\item{course_instance_id}{Integer value associated with the course.}
}
\value{
A \code{data.frame}
}
\description{
Obtains the course gradebook
}
