% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pli-httr.R
\name{pl_get}
\alias{pl_get}
\title{Perform a GET call to the PrairieLearn API}
\usage{
pl_get(url, args = NULL)
}
\arguments{
\item{url}{URL with endpoint}

\item{args}{Values to pass onto \code{query} component.}
}
\description{
Calls into the specified PrairieLearn endpoint with a \code{GET} request.
}
\details{
Structures the API call as:\if{html}{\out{<div class="bash">}}\preformatted{curl -H "Private-Token: TOKEN" https://prairielearn.engr.illinois.edu/pl/api/v1/<REST_OF_PATH>
}\if{html}{\out{</div>}}
}
\keyword{internal}
