% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pli-package.R
\docType{package}
\name{pli-package}
\alias{pli}
\alias{pli-package}
\title{pli: Access PrairieLearn's Web API from R}
\description{
Provides a set of functions that pulls data from PrairieLearns' 
    Web API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/illinois-r/pli}
  \item Report bugs at \url{https://github.com/illinois-r/pli/issues}
}

}
\author{
\strong{Maintainer}: James Joseph Balamuta \email{balamut2@illinois.edu} (\href{https://orcid.org/0000-0003-2826-8458}{ORCID}) [copyright holder]

}
\keyword{internal}
